feat: 🎸 Npm and npx is now working
Is probably very unstable and buggy rn, I will work to improve it

+++ src/filesystem/node.rs
@@ -2,34 +2,31 @@ use std::path::PathBuf;
 use crate::commands::NodeVersion;
 use super::Error;
 
-/// This function will create the directory for specific version of node.js
-/// and put all executable files into it.
+/// This function will download the compressed node.js zip file
+/// and extract all of it's content to the specified path
 /// 
 /// # Arguments:
 /// 
 /// * `data` - Data of the node version we want to install
 /// * `cnvmpath` - Path to the cnvm folder
 /// * `node_bytes` - Bytes of the node executable
-/// * `npm_bytes` - Bytes of the npm executable 
 pub fn create_node(
     data: NodeVersion,
     path: &PathBuf, 
-    node_bytes: &Vec<u8>,
-    npm_bytes: &Vec<u8>
+    node_bytes: &Vec<u8>
 ) -> Result<(), Error> {
-    let node_path = path.join(data.version);
-    
-    std::fs::create_dir_all(&node_path).map_err(|err| {
-        Error::PermissionError(Some(err.to_string()))
-    })?;
-    
-    std::fs::write(&node_path.join("node.exe"), node_bytes).map_err(|err| {
+    // Create a zip file from the node bytes
+    let mut node_zip = zip::ZipArchive::new(
+        std::io::Cursor::new(node_bytes)
+    ).unwrap();
+
+    // Create the directory for the node version
+    std::fs::create_dir_all(&path).map_err(|err| {
         Error::PermissionError(Some(err.to_string()))
     })?;
 
-    std::fs::write(&node_path.join("npm"), npm_bytes).map_err(|err| {
-        Error::PermissionError(Some(err.to_string()))
-    })?;
+    // Extract the node executable from the zip file
+    node_zip.extract(path).unwrap();
 
     Ok(())
 }
@@ -66,9 +63,11 @@ pub fn symlink_node(cnvmpath: &PathBuf, nodepath: &PathBuf) -> Result<(), Error>
 /// 
 /// * `Result<(), Error>` - Result of the function
 pub fn remove_symlink(cnvmpath: &PathBuf, nodepath: &PathBuf) -> Result<(), Error> {
-    std::fs::remove_dir_all(nodepath).map_err(|err| {
-        Error::PermissionError(Some(err.to_string()))
-    })?;
+    if nodepath.exists() {
+        std::fs::remove_dir_all(nodepath).map_err(|err| {
+            Error::PermissionError(Some(err.to_string()))
+        })?;
+    }
 
     Ok(())
 }
